requirements.txt
absl-py==2.4.0
accelerate==1.13.0
addict==2.4.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astor==0.8.1
asttokens==3.0.1
async-lru==2.3.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.1.0
auto_gptq==0.7.1
autoflake==2.3.1
av==12.0.0
azure-core==1.39.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.23.0
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bitsandbytes==0.43.1
black==24.3.0
bleach==6.3.0
blinker==1.9.0
boto3==1.42.73
botocore==1.42.73
braceexpand==0.1.7
calmsize==0.1.3
ccimport==0.4.4
cffi==2.0.0
circuitbreaker==2.1.3
click==8.3.1
cloudpickle==3.1.2
colorama
coloredlogs==15.0.1
comm==0.2.3
conda-pack==0.7.1
ConfigArgParse==1.7.5
contourpy==1.3.2
cramjam==2.11.0
crcmod==1.7
cryptography==46.0.5
cuda-python==12.4.0
cumm-cu121==0.7.11
cycler==0.12.1
cyclopts==4.10.0
dash==4.0.0
dataclasses==0.6
dataclasses-json==0.6.7
datasets==4.8.3
debugpy==1.8.20
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
diffusers==0.35.1
dill==0.4.1
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22.4
easydict==1.13
editables==0.5
einops==0.8.2
einops-exts==0.0.4
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.2.1
fastapi==0.135.1
fastavro==1.9.4
fasteners==0.19
fastjsonschema==2.21.2
ffmpy==1.0.0
fire==0.7.1
flake8==7.0.0
Flask==3.0.3
fonttools==4.62.1
fqdn==1.5.1
freetype-py==2.5.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gekko==1.3.2
gitdb==4.0.12
GitPython==3.1.46
glcontext==3.0.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-storage==2.10.0
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
gradio==5.49.0
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.78.0
gsplat==1.5.3
h11==0.16.0
h5py==3.12.1
hatch-requirements-txt==0.4.1
hatchling==1.29.0
hdfs==2.7.3
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
igraph==0.11.8
ImageIO==2.37.3
imath==0.0.2
importlib-metadata==6.11.0
iniconfig==2.3.0
inquirerpy==0.3.4
iopath==0.1.10
ipycanvas==0.14.3
ipyevents==2.0.4
ipykernel==6.29.5
ipython==8.38.0
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.3.7
jedi==0.19.2
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonlines==4.0.0
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==7.4.9
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaolin==0.17.0
kiwisolver==1.5.0
lark==1.3.1
lazy-loader==0.5
libcst==1.8.6
librosa==0.10.1
lightning==2.3.3
llvmlite==0.46.0
loguru==0.7.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
mock==4.0.3
moderngl==5.12.0
moreorless==0.5.0
mosaicml-streaming==0.7.5
msal==1.35.1
msal-extensions==1.3.1
msgpack==1.1.2
multiprocess==0.70.19
mypy_extensions==1.1.0
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.13.0
notebook==7.5.5
notebook_shim==0.2.4
numba==0.64.0
numpy==1.26.4
nvidia-cuda-nvcc-cu12==12.4.131
nvidia-ml-py==13.595.45
nvidia-pyindex==1.0.9
objsize==0.7.0
OCCUtils==0.1.dev0
oci==2.168.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.9.0.80
OpenEXR==3.3.3
optimum==2.1.0
optree==0.14.1
orjson==3.10.0
overrides==7.7.0
packaging==24.2
Panda3D==1.10.16
panda3d-gltf==1.2.1
panda3d-simplepbr==0.13.1
pandas==2.3.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.6
pathos==0.3.5
pathspec==1.0.4
pccm==0.4.16
pdoc3==0.10.0
peft==0.18.1
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
pip-system-certs==4.0
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
plyfile==1.1.3
point-cloud-utils==0.29.5
polyscope==2.3.0
pooch==1.9.0
portalocker==3.2.0
pox==0.3.7
ppft==1.7.8
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybind11==3.0.2
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==3.0
pydantic==2.11.10
pydantic_core==2.33.2
pydot==1.4.2
pydub==0.25.1
pyflakes==3.2.0
pyglet==2.1.13
pygltflib==1.16.5
Pygments==2.19.2
PyJWT==2.12.1
pymeshfix==0.17.0
pymongo==4.6.3
PyNaCl==1.6.2
pynvml==13.0.1
PyOpenGL==3.1.0
pyOpenSSL==25.3.0
pyquaternion==0.9.9
pyrender==0.1.45
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
python-pycg==0.9.2
python-snappy==0.7.3
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
pytz==2026.1.post1
pyvista==0.47.1
pyzmq==27.1.0
randomname==0.2.1
referencing==0.37.0
regex==2026.2.28
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rich-rst==1.3.2
roma==1.5.1
rootutils==1.0.7
rouge==1.0.1
rpds-py==0.30.0
Rtree==1.3.0
ruff==0.15.7
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
sagemaker==2.242.0
sagemaker-core==1.0.77
schema==0.7.8
scikit-image==0.23.1
scikit-learn==1.7.2
scipy==1.15.3
scooby==0.11.0
screeninfo==0.8.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==2.1.0
sentence-transformers==2.6.1
sentencepiece==0.2.1
sentry-sdk==2.55.0
setproctitle==1.3.7
shellingham==1.5.4
simplejson==3.19.2
six==1.17.0
smdebug-rulesconfig==1.0.1
smmap==5.0.3
smplx==0.1.28
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spconv-cu121==2.3.8
stack-data==0.6.3
starlette==0.52.1
stdlibs==2026.2.26
submitit==1.5.4
sympy==1.13.1
tabulate==0.10.0
tblib==3.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==3.3.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.9.16
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.3
torch==2.5.1
torchvision==0.20.1
tornado==6.5.5
trailrunner==1.4.0
traitlets==5.14.3
transformers==4.46.0
trimesh==4.11.4
triton==3.1.0
trove-classifiers==2026.1.14.14
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
usd-core==26.3
usort==1.0.8.post1
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@3913c65d81e05e47b9f367250cf8c0f7462a0900
uvicorn==0.42.0
vtk==9.6.0
wadler_lindig==0.1.7
wandb==0.20.0
warp-lang==1.12.0
wcwidth==0.2.14
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.15
wrapt==2.1.2
xatlas==0.0.9
xformers==0.0.28.post3
xxhash==3.6.0
yacs==0.1.8
zipp==3.23.0
zstd==1.5.7.3
